Never trust user input

+++ newfile/delete.go
@@ -11,10 +11,19 @@ func DeleteHandler(c *fiber.Ctx) error {
 	id := c.Params("id")
 	fileName := c.Params("filename")
 
+	if fileName == "" || id == "" {
+		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
+			"message": "Please provide a file id and filename",
+			"error":   nil,
+			"delete":  false,
+		})
+	}
+
 	if err := os.Remove("tmp/" + id + "/" + fileName); err != nil {
 		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
 			"message": "file delete error",
 			"error":   err.Error(),
+			"delete":  false,
 		})
 	}
 
@@ -23,10 +32,14 @@ func DeleteHandler(c *fiber.Ctx) error {
 		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
 			"message": "db delete error",
 			"error":   err.Error(),
+			"delete":  false,
 		})
 	}
 
 	return c.Status(fiber.StatusOK).JSON(fiber.Map{
 		"message": "File deleted successfully",
+		"error":   nil,
+		"delete":  true,
 	})
+
 }
